Добавлена пагинация, фильтры и SSO-авторизация
В контроллеры сравнения добавлены параметры пагинации, фильтрации и поиска, реализована поддержка этих опций в Query/Handler и расширена статистика в DTO. Введён новый контроллер SsoAuthController с авторизацией через SSO, определением ролей и выдачей JWT. Везде double заменён на float для корректной работы с MSSQL. Обновлены SQL-скрипты и строки подключения для локальных MSSQL-баз. Проведён рефакторинг и унификация логики фильтрации и поиска.

+++ AccountingScholarships.Domain/Entities/Real/epvosso/SpecialitiesEpvo.cs
@@ -8,17 +8,17 @@ namespace AccountingScholarships.Domain.Entities.Real.epvosso
 {
     public class SpecialitiesEpvo
     {
-        public double? UniversityId { get; set; }
-        public double? Id { get; set; }
-        public double? profCafId { get; set; }
+        public float? UniversityId { get; set; }
+        public float? Id { get; set; }
+        public float? profCafId { get; set; }
         public string? NameRu { get; set; }
         public string? NameKz { get; set; }
         public string? NameEn { get; set; }
         public bool? DoubleDiploma { get; set; }
         public bool? JointEp { get; set; }
         public string? SpecializationCode { get; set; }
-        public double? StatusEp { get; set; }
-        public double? EduProgType { get; set; }
+        public float? StatusEp { get; set; }
+        public float? EduProgType { get; set; }
         public bool? Default { get; set; }
         public bool? Interdisciplinary { get; set; }
         public bool? EducationProgram { get; set; }
